Rename packages from prepper to processor
