+++ data-prepper-plugins/common/src/test/java/org/opensearch/dataprepper/plugins/processor/state/ProcessorStateTest.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package org.opensearch.dataprepper.plugins.prepper.state;
+package org.opensearch.dataprepper.plugins.processor.state;
 
 import org.opensearch.dataprepper.processor.state.ProcessorState;
 import org.junit.After;
